requirements.txt
PyYAML==6.0.1
numpy==1.25.2
matplotlib==3.7.2
scipy==1.11.2
scikit-learn==1.3.0